test: enable DragAndDrop-integ-test
