+++ test/UnitTestSuite.js
@@ -82,7 +82,7 @@ define(function (require, exports, module) {
     require("spec/CSSInlineEdit-integ-test");
     require("spec/DocumentCommandHandlers-integ-test");
     require("spec/DocumentManager-integ-test");
-    // require("spec/DragAndDrop-test");
+    require("spec/DragAndDrop-integ-test");
     // require("spec/EditorOptionHandlers-test");
     // require("spec/ExtensionInstallation-test");
     // require("spec/ExtensionUtils-test");
